% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manoby.R
\name{manoby}
\alias{manoby}
\title{Interactive panel MANOVA classic, non parametric and bayesian}
\usage{
manoby(dataset = FALSE)
}
\arguments{
\item{dataset}{Data set}
}
\value{
A shiny panel with the multivariate analyze of variance, PERMANOVA and Bayesian analyzes of variance, based on selection the dependents variables and independent variable of the data set provided, presents decision diagram that suggets correct method use.
}
\description{
Interactive panel to visualize and develop one-way multivariate analysis of variance models, from the classical, non-parametric and Bayesian approach.
}
\examples{
\donttest{
library(MANOVA.RM)
data(EEGwide)
manoby(EEGwide)
}
}
