% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapeBay-package.R
\docType{package}
\name{MapeBay-package}
\alias{MapeBay-package}
\alias{MapeBay}
\title{The 'MapeBay' package.}
\description{
Package developed for the visualization and presentation of one-way multivariate analysis of variance models, with a classical, non-parametric and Bayesian approach.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
